requirements.txt
# Core API
fastapi>=0.115.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
pydantic>=2.8.0
pydantic-settings==2.2.1

# LangChain + LLMs
langchain==0.2.1
langchain-community==0.2.1
langchain-anthropic==0.1.15

# Embeddings & Vector DBs
voyageai==0.3.2
langchain-voyageai==0.1.1
chromadb==0.5.0
faiss-cpu==1.8.0

# PDF processing
pypdf==4.2.0
langchain-community[pdf]==0.2.1

# GCS
google-cloud-storage==2.16.0

# Keyword search
rank-bm25==0.2.2

# MCP
mcp==1.0.0                    # Model Context Protocol SDK
anthropic[mcp]==0.40.0        # Anthropic SDK with MCP support

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.1
